[FsxS3Access] Add missing permissions needed to import/export from S3 when using Fsx for Lustre via API
Creation of a cluster with FsxLisuter storage via ParallelCluster API fails due to the user role missing some permissions needed to import/export from S3.
This commit creates an inline-policy with the needed permissions and associates it with the user role.
The inline-policy is only created if the API stack is created with certain parameters (EnableFSxS3Access and FsxS3Buckets).